Update crc-org/machine module
With this update type of memory and cpu is changed to uint from int
which will allow us to fix some of interter overflow issue with latest
golint ci.

+++ pkg/drivers/vfkit/driver_darwin.go
@@ -173,7 +173,7 @@ func (d *Driver) Start() error {
 	bootLoader := config.NewEFIBootloader(efiStore, create)
 
 	vm := config.NewVirtualMachine(
-		uint(d.CPU),
+		d.CPU,
 		uint64(d.Memory),
 		bootLoader,
 	)
